backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
python-dotenv==1.0.0
Markdown==3.8
sqlparse==0.5.3
tzdata==2025.2


absl-py==2.1.0
addict==2.4.0
annotated-types==0.7.0
anyio==4.7.0
basicsr==1.3.4.9
certifi==2024.12.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.8
contourpy==1.2.1
cycler==0.12.1
djangorestframework_simplejwt==5.5.0
docker-pycreds==0.4.0
einops==0.8.1
filelock==3.18.0
fonttools==4.53.0
fsspec==2025.3.2
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
h11==0.14.0
h5py==3.12.1
-e git+https://github.com/XPixelGroup/HAT.git@1638a9a822581657811867bf670717f8371fc3e5#egg=hat
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.6.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.0
onnx==1.17.0
opencv-python==4.10.0.84
optree==0.13.1
packaging==24.1
pandas==2.2.2
pillow==10.3.0
piq
platformdirs==4.3.8
protobuf==6.30.2
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pygame==2.6.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-telegram-bot==21.10
pyttsx3==2.98
pytz==2024.1
pywin32==308
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-image==0.25.2
scikit-learn==1.6.0
scipy==1.14.1
sentry-sdk==2.27.0
setproctitle==1.3.6
setuptools==80.3.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
tb-nightly==2.20.0a20250601
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2025.5.26
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.4.0
wandb==0.19.11
Werkzeug==3.1.3
yapf==0.43.0

